instanceNameSuffix func moved to utils
